Update/ Restructure files & Fix/ logic refreshtoken + userService

+++ micro-be/IdentityService/src/main/java/com/microservice/IdentityService/Application/Services/UserService.java
@@ -1,6 +1,7 @@
 package com.microservice.IdentityService.Application.Services;
 
 
+import com.microservice.IdentityService.Application.Abstrations.IUserService;
 import com.microservice.IdentityService.Application.Dtos.User.Request.CreateUserRequest;
 import com.microservice.IdentityService.Application.Dtos.User.Request.UserCommonRequest;
 import com.microservice.IdentityService.Application.Dtos.User.Respone.UserResponse;
@@ -10,7 +11,6 @@ import com.microservice.IdentityService.Application.Persistences.Repositories.IU
 import com.microservice.IdentityService.Domain.Entities.Role;
 import com.microservice.IdentityService.Domain.Entities.User;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,14 +18,13 @@ import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
-public class UserService {
+public class UserService implements IUserService {
 
     private final IUserRepository userRepository;
     private final IRoleRepository roleRepository;
-    private final PasswordEncoder passwordEncoder;
     private final UserProfile userMapper;
 
-
+    @Override
     public UserResponse createUser(CreateUserRequest request) {
 
         UserCommonRequest userCommonRequest = request.getUserCommonRequest();
@@ -34,42 +33,30 @@ public class UserService {
             throw new RuntimeException("Email already exists");
         }
 
-        User user = new User();
-        user.setFirstName(userCommonRequest.getFirstName());
-        user.setLastName(userCommonRequest.getLastName());
-        user.setEmail(userCommonRequest.getEmail());
-        user.setUsername(userCommonRequest.getUsername());
-        user.setPassword(passwordEncoder.encode(request.getPassword()));
-        user.setPhone(userCommonRequest.getPhone());
-        user.setAddress(userCommonRequest.getAddress());
-        user.setGender(userCommonRequest.getGender());
-        user.setDateOfBirth(userCommonRequest.getDateOfBirth());
-        user.setAvatarUrl(userCommonRequest.getAvatarUrl());
-        user.setAvatarPublicId(userCommonRequest.getAvatarPublicId());
-
+        User user = userMapper.fromCreateRequest(request);
         if (userCommonRequest.getRoleId() != null) {
-            Role role = new Role();
-            role.setId(UUID.fromString(userCommonRequest.getRoleId()));
+            Role role = roleRepository.findById(UUID.fromString(userCommonRequest.getRoleId()))
+                    .orElseThrow(() -> new RuntimeException("Role not found"));
             user.setRole(role);
         }
         User savedUser = userRepository.save(user);
         return userMapper.toResponse(savedUser);
     }
-
+    @Override
     public UserResponse getById(UUID id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("User not found"));
 
         return userMapper.toResponse(user);
     }
-
+    @Override
     public List<UserResponse> getAll() {
         return userRepository.findAll()
                 .stream()
                 .map(userMapper::toResponse)
                 .toList();
     }
-
+    @Override
     public UserResponse update(UserCommonRequest request) {
         User user = userRepository.findByEmail(request.getEmail())
                 .orElseThrow(() -> new RuntimeException("User not found"));
@@ -82,8 +69,11 @@ public class UserService {
         userRepository.save(user);
         return userMapper.toResponse(user);
     }
-
+    @Override
     public void deleteById(UUID id) {
+        if (!userRepository.existsById(id)) {
+            throw new RuntimeException("User not found");
+        }
         userRepository.deleteById(id);
     }
 }
